fix for mingw compilation, +don't waste time on realloc if the size is the same

+++ sources/src/data.cpp
@@ -42,6 +42,6 @@ int move_nodes;
 int search_depth;
 int start_time;
 ENTRY *tt;
-size_t tt_size;
-size_t tt_mask;
+unsigned int tt_size;
+unsigned int tt_mask;
 int tt_date;
